# services/api/internal/openapi/openapi.yaml
openapi: "3.0.0"
info:
  title: AITC API
  version: "1.0.0"
  description: |
    Autonomous Identity Threat Containment (AITC) REST API.
    All endpoints require a valid Bearer token (Okta OIDC).
    RBAC roles: Admin, Analyst, Auditor, ReadOnly.
servers:
  - url: http://localhost:8085
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
    Case:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
        status:
          type: string
          enum: [open, investigating, resolved, closed]
        severity:
          type: string
          enum: [critical, high, medium, low]
        summary:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CaseList:
      allOf:
        - $ref: '#/components/schemas/Pagination'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Case'
...
    EnableAutopilotRequest:
      type: object
      required: [confirm]
      properties:
        confirm:
          type: boolean
          description: Must be explicitly true to enable autopilot.
    EnableAutopilotResponse:
      type: object
      properties:
        autopilot_enabled:
          type: boolean
        tenant_id:
          type: string
...
security:
  - bearerAuth: []
paths:
  /v1/cases:
    get:
      summary: List cases
      description: |
        Returns a paginated list of threat cases for the authenticated tenant.
        **Required role:** ReadOnly or higher (ReadOnly, Analyst, Auditor, Admin).
      operationId: listCases
      parameters:
        - name: status
          in: query
          schema:
            type: string
          description: Filter by case status (open, investigating, resolved, closed)
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Paginated list of cases
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseList'
        "400":
          description: Bad request (missing tenant_id)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden (insufficient role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/cases/{id}:
    get:
      summary: Get a case by ID
      description: |
        Returns a single case by ID.
        **Required role:** ReadOnly or higher.
      operationId: getCase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Case details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Case not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v1/autopilot/enable:
    post:
      summary: Enable autopilot
      description: |
        Enables autopilot for the tenant. Requires ≥7 days since onboarding.
        The request body must contain `{"confirm": true}`.
        **Required role:** Admin only.
      operationId: enableAutopilot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnableAutopilotRequest'
      responses:
        "200":
          description: Autopilot enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnableAutopilotResponse'
        "400":
          description: Bad request (confirm not true, or tenant not found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden (Admin role required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "409":
          description: Learning mode not complete (< 7 days since onboarding)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...